[V2] Migrate all builders to hashicorp/go-azure-sdk  (#326)
* ARM Builder

Rebasing entire ARM builder migration on top of main

Remove some missed old SDK stuff, including the autorest To package from ARM Builder

Update oidc-example.json

Pass SIG ID subscription through to state

Use Hamilton MSGraph SDK to fetch Object ID for Windows builds

Format and add comment

commit generated code

Switch AuthType back to private variable and use helper function to get it, switch from MeClient to ServicePrincipals client to get objectID

Print MSgraph error again

Don't use Hamilton, get it from the Access Token

fmt and pass in correct context on cleanup

Mess with backoff time while testing to see if this can be a bit smoother of a process for users

Don't use context.TODOs in ARM Builder client calls

Use hcl file in oidc test

Use Giovanni instead of Blob Container serivce

Implement new SDK environment logic

Address fmt

Remove unused function, go mod tidy

Switch to forked release process

* Azure Germany was shutdown in 2021

* Remove Template Capture from ARM Builder, this work will need to be replicated to DTL/Chroot

* Fail fast: Check if a SIG Destination Image Version exists before trying to create it

Implement Lucas's feedback

Make generate

* Port DTL Builder to new SDK

=======

DTL almost done! Just need to get VMID instead of capture template method and start testing

Port DTL provisioner to new SDK

Make generate

Accidently commited playing around with this field, but was missing comma

Appease the lint lord

DTL never supported VHDs, it just had all the code of it, so I didn't need to port some of this

Use retry config instead of an infinite loop

Delete more unused code

Update comments

Address linter

* Port CHRoot Builder and rename hashi* imports

====
EOD WIP: I've ported all the client code over, on Monday I will test that it works in a VM before porting the unit tests as the mocking strategy used is heavily reliant on auto rest, and I want to port it to stretchify mocks that don't care as much about the autorest specifics

Don't use Subscription ID as Snapshot URL

Large commit: Chroot unit tests passing, azure-sdk-for-go no longer imported by go.mod, remove references to device code, clear up missing pieces of DTL builder

Fix linux disk attatcher

Remove unnused mock that I added earlier

Re-add Fetching of Tenant ID

Forgot to use New Steps

null check was wrong way for SDK model

see above

Don't try and get tenant ID if we're using CLI auth

Address linter, fix linux test

StepCreateNewDiskTest -> Return proper autorest future fake

remove device code acceptance tests, handle flakey unit test by sorting resulting objects to insure consistant comparison

Update builder/azure/chroot/step_verify_source_disk_test.go


Remove CHRoot Acceptance Test && rename all hashi* imports removing prefix

Make generate and fix lint

Readd subscription, tenant ID fetching for CLI auth, so we know which subscription to save stuff to, this caused acceptance test failures which this commit fixes

Lint

* Add WinRM secret expiration, Fast Fail on invalid SIG version, Use context timeout throughout codebase

Reset readme, version file, and workflows to state in main origin

* Pass in expiry time when making a new Key Vault as well, lint

[chroot] PollingDelay != Polling Duration >_>

* Fix rebasing error where we were logging for skip deletion always and remove VHD Deprecation

* Update goreleaser and codeowners to remove jenna's fork logic

Okay linter..

* Packer Plugin SDK v0.5.1

* Partially implement Wilken's feedback

* Add comments based on Wilken's feedback, don't panic on unexpected AuthType error

* Implement Wilken's feedback on Null Model SDK Error

* Implement Lucas's feeback

+++ builder/azure/chroot/diskattacher_other.go
@@ -10,6 +10,6 @@ import (
 	"context"
 )
 
-func (da diskAttacher) WaitForDevice(ctx context.Context, lun int32) (device string, err error) {
+func (da diskAttacher) WaitForDevice(ctx context.Context, lun int64) (device string, err error) {
 	panic("The azure-chroot builder does not work on this platform.")
 }
